+++ trixnity-messenger-compose-view/karma.config.d/karma-timeout.js
@@ -1,5 +1,5 @@
 if (!!config.client) {
     config.client.mocha = config.client.mocha || {}
-    config.client.mocha.timeout = 30000
+    config.client.mocha.timeout = 300000
 }
-config.browserNoActivityTimeout = 30000
+config.browserNoActivityTimeout = 300000