+++ trixnity-messenger-compose-view/src/androidUnitTest/kotlin/de/connect2x/trixnity/messenger/compose/view/util/platformHttpEngine.android.kt
@@ -1,0 +1,7 @@
+package de.connect2x.trixnity.messenger.compose.view.util
+
+import io.ktor.client.engine.HttpClientEngine
+
+actual fun platformHttpEngine(): HttpClientEngine {
+    throw Exception("tests requiring a MatrixMultiMessenger should be run as instrumented tests")
+}