+++ trixnity-messenger-compose-view/src/androidUnitTest/kotlin/de/connect2x/trixnity/messenger/compose/view/util/platformScreenshot.android.kt
@@ -1,0 +1,7 @@
+package de.connect2x.trixnity.messenger.compose.view.util
+
+import androidx.compose.ui.test.SemanticsNodeInteraction
+
+actual suspend fun SemanticsNodeInteraction.screenshot(path: String) {
+    throw Exception("tests requiring a MatrixMultiMessenger should be run as instrumented tests")
+}