+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/settings/AccountSetupWizard.kt
@@ -124,7 +124,7 @@ fun AccountSetupWizard(showAccountBootstrapWrapper: Wrapper.ShowAccountSetup) {
             }
         }
     }
-    Wizard(wizardSteps, true)
+    Wizard(wizardSteps, useDefaultBackHandler = true, wizardId = "AccountSetupWizard")
 }
 
 private fun wizardStepExplanation(