+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/verification/DeviceVerificationSwitch.kt
@@ -150,7 +150,7 @@ fun BoxScope.DeviceVerificationWizardStepSwitch(
             }
         }
     )
-    Wizard(listOf(step))
+    Wizard(listOf(step), wizardId = "DeviceVerificationWizardStep")
 }