+++ trixnity-messenger-compose-view/src/commonTest/kotlin/de/connect2x/trixnity/messenger/compose/view/util/platformName.kt
@@ -1,0 +1,3 @@
+package de.connect2x.trixnity.messenger.compose.view.util
+
+expect val platformName: String