+++ trixnity-messenger-compose-view/src/wasmJsTest/kotlin/de/connect2x/trixnity/messenger/compose/view/util/platformName.wasm.kt
@@ -1,0 +1,3 @@
+package de.connect2x.trixnity.messenger.compose.view.util
+
+actual val platformName: String = "WASM"