+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- SDK: add an abi checker to the CI
 - Room list context menu now includes option to mark a room as read
 
 ### Changed